+++ addon/components/paper-form.js
@@ -14,16 +14,35 @@ const { Component, computed } = Ember;
  */
 export default Component.extend(ParentMixin, {
   layout,
-  tagName: '',
+  tagName: 'form',
+
+  inputComponent: 'paper-input',
+  submitButtonComponent: 'paper-button',
+  selectComponent: 'paper-select',
+  autocompleteComponent: 'paper-autocomplete',
+
   isValid: computed.not('isInvalid'),
   isInvalid: computed('childComponents.@each.isInvalid', function() {
     return this.get('childComponents').isAny('isInvalid');
   }),
+
+  isTouched: computed('childComponents.@each.isTouched', function() {
+    return this.get('childComponents').isAny('isTouched');
+  }),
+
+  isInvalidAndTouched: computed.and('isInvalid', 'isTouched'),
+
+  submit() {
+    this.send('onSubmit');
+    return false;
+  },
+
   actions: {
     onValidityChange() {
-      if (this.get('lastIsValid') !== this.get('isValid')) {
-        this.sendAction('onValidityChange', this.get('isValid'));
+      if (this.get('lastIsValid') !== this.get('isValid') || this.get('lastIsTouched') !== this.get('isTouched')) {
+        this.sendAction('onValidityChange', this.get('isValid'), this.get('isTouched'), this.get('isInvalidAndTouched'));
         this.set('lastIsValid', this.get('isValid'));
+        this.set('lastIsTouched', this.get('isTouched'));
       }
     },
     onSubmit() {