+++ addon/components/paper-input.js
@@ -25,8 +25,8 @@ export default Component.extend(FocusableMixin, ColorMixin, ChildMixin, Validati
   classNameBindings: [
     'hasValue:md-input-has-value',
     'isInvalidAndTouched:md-input-invalid',
-    'eitherIcon:md-has-icon',
-    'iconRight:md-icon-right',
+    'hasLeftIcon:md-icon-left',
+    'hasRightIcon:md-icon-right',
     'focused:md-input-focused',
     'block:md-block'
   ],
@@ -35,7 +35,12 @@ export default Component.extend(FocusableMixin, ColorMixin, ChildMixin, Validati
   tabindex: null,
   hideAllMessages: false,
   isTouched: false,
-  isInvalid: computed.or('validationErrorMessages.length', 'isNativeInvalid'),
+
+  iconComponent: 'paper-icon',
+
+  // override validation mixin `isInvalid` to account for the native input validity
+  isInvalid: computed.or('hasErrorMessages', 'isNativeInvalid'),
+
   hasValue: computed('value', 'isNativeInvalid', function() {
     let value = this.get('value');
     let isNativeInvalid = this.get('isNativeInvalid');
@@ -51,7 +56,8 @@ export default Component.extend(FocusableMixin, ColorMixin, ChildMixin, Validati
     return `${currentLength}/${this.get('maxlength')}`;
   }),
 
-  eitherIcon: computed.or('icon', 'iconRight'),
+  hasLeftIcon: computed.bool('icon'),
+  hasRightIcon: computed.bool('iconRight'),
   isInvalidAndTouched: computed.and('isInvalid', 'isTouched'),
 
   validationProperty: 'value', // property that validations should be run on