+++ addon/components/paper-item.js
@@ -20,7 +20,10 @@ export default Component.extend(RippleMixin, ParentMixin, {
 
   // Ripple Overrides
   rippleContainerSelector: '.md-no-style',
-  noink: computed.not('shouldBeClickable'),
+  // disable ripple when we have a primary action or when we don't have a proxied component
+  noink: computed('hasPrimaryAction', 'hasProxiedComponent', function() {
+    return this.get('hasPrimaryAction') || !this.get('hasProxiedComponent');
+  }),
 
   center: false,
   dimBackground: true,
@@ -28,7 +31,7 @@ export default Component.extend(RippleMixin, ParentMixin, {
 
   classNameBindings: [
     'hasProxiedComponent:md-proxy-focus', 'shouldBeClickable:md-clickable',
-    'focused:md-focused'
+    'focused:md-focused', 'hasPrimaryAction:_md-button-wrap'
   ],
   attributeBindings: ['role', 'tabindex'],
   role: 'listitem',
@@ -58,5 +61,13 @@ export default Component.extend(RippleMixin, ParentMixin, {
         component.processProxy();
       }
     });
+  },
+
+  mouseEnter(ev) {
+    this.sendAction('onMouseEnter', ev);
+  },
+
+  mouseLeave(ev) {
+    this.sendAction('onMouseLeave', ev);
   }
 });