+++ app/utils/clamp.js
@@ -1,0 +1,1 @@
+export { default } from 'ember-paper/utils/clamp';