+++ tests/integration/components/paper-dialog-test.js
@@ -9,11 +9,10 @@ moduleForComponent('paper-dialog', 'Integration | Component | paper dialog', {
 
 test('should render proper dialog wrapping selectors', function(assert) {
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{paper-dialog}}
   `);
 
-  let selectors = this.$('#paper-wormhole .md-dialog-container md-dialog');
+  let selectors = $('#ember-testing > .md-dialog-container md-dialog');
 
   assert.ok(selectors.length, 'has proper selector nesting');
 
@@ -21,11 +20,9 @@ test('should render proper dialog wrapping selectors', function(assert) {
 
 test('should render empty dialog when blockless', function(assert) {
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{paper-dialog}}
   `);
-
-  let dialogContent = this.$().find('md-dialog').html()
+  let dialogContent = $('md-dialog').html()
     .replace('<!--', '').replace('-->', '').trim();
 
   assert.equal(dialogContent, '', 'has an empty dialog container');
@@ -33,24 +30,22 @@ test('should render empty dialog when blockless', function(assert) {
 
 test('should yield content as a block component', function(assert) {
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{#paper-dialog}}
       Lorem ipsum.
     {{/paper-dialog}}
   `);
 
-  let dialogContent = this.$().find('md-dialog').html().trim();
+  let dialogContent = $('md-dialog').html().trim();
 
   assert.equal(dialogContent, 'Lorem ipsum.', 'yielded dialog content');
 });
 
-test('should render in default wormhole if no parent is defined', function(assert) {
+test('should render in ember-testing if no parent is defined', function(assert) {
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{paper-dialog}}
   `);
 
-  assert.ok(this.$().find('#paper-wormhole md-dialog'), 'rendered in default');
+  assert.ok($().find('#ember-testing md-dialog'), 'rendered in default');
 });
 
 test('should render in specific wormhole if parent is defined', function(assert) {
@@ -62,8 +57,9 @@ test('should render in specific wormhole if parent is defined', function(assert)
     {{/paper-dialog}}
   `);
 
-  assert.ok(!this.$('#paper-wormhole md-dialog').length, 'did not render in default');
-  assert.ok(this.$('#sagittarius-a md-dialog').length, 'rendered in parent');
+  assert.ok(!$('#paper-wormhole md-dialog').length, 'did not render in default');
+  assert.ok($('#sagittarius-a md-dialog').length, 'rendered in parent');
+  assert.ok($('#ember-testing md-dialog').length, 'still rendered in ember-testing');
 });
 
 test('should only prevent scrolling behind scoped modal', function(assert) {
@@ -72,25 +68,41 @@ test('should only prevent scrolling behind scoped modal', function(assert) {
     {{paper-dialog parent="#sagittarius-a"}}
   `);
 
-  assert.equal(this.$('md-backdrop').css('position'), 'absolute', 'backdrop is absolute');
+  assert.equal($('md-backdrop').css('position'), 'absolute', 'backdrop is absolute');
+});
+
+test('backdrop is opaque by default', function(assert) {
+  this.render(hbs`
+    <div id="paper-wormhole"></div>
+    {{paper-dialog}}
+  `);
+
+  assert.ok($('md-backdrop').hasClass('md-opaque'), 'backdrop is opaque');
+});
+
+test('backdrop opaqueness can be disabled ', function(assert) {
+  this.render(hbs`
+    <div id="paper-wormhole"></div>
+    {{paper-dialog opaque=false}}
+  `);
+
+  assert.notOk($('md-backdrop').hasClass('md-opaque'), 'backdrop is not opaque');
 });
 
 test('should prevent scrolling entirely behind fixed modal', function(assert) {
   assert.expect(1);
 
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{paper-dialog}}
   `);
 
   assert.equal(
-    this.$('md-backdrop').css('position'), 'fixed', 'backdrop is fixed'
+    $('md-backdrop').css('position'), 'fixed', 'backdrop is fixed'
   );
 });
 
 test('applies transitions when opening and closing', function(assert) {
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{#if dialogOpen}}
       {{paper-dialog}}
     {{/if}}
@@ -98,7 +110,7 @@ test('applies transitions when opening and closing', function(assert) {
   this.set('dialogOpen', true);
 
   let getDialogTransform = () => {
-    let dialog = this.$('md-dialog').get(0);
+    let dialog = $('md-dialog').get(0);
     assert.ok(dialog, 'dialog found');
     return dialog && (dialog.style.webkitTransform || dialog.style.transform);
   };
@@ -128,28 +140,26 @@ test('click outside should close dialog if clickOutsideToClose', function(assert
   });
 
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{#if dialogOpen}}
       {{paper-dialog clickOutsideToClose=true onClose=closeDialog}}
     {{/if}}
   `);
 
-  assert.ok(this.$('md-dialog').length, 'dialog is showing');
+  assert.ok($('md-dialog').length, 'dialog is showing');
 
-  this.$('.md-dialog-container').mousedown().mouseup().click();
+  $('.md-dialog-container').mousedown().mouseup().click();
 });
 
 test('click outside should not close dialog by default', function(assert) {
   assert.expect(2);
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{paper-dialog}}
   `);
 
-  assert.ok(this.$('md-dialog').length, 'dialog is showing');
+  assert.ok($('md-dialog').length, 'dialog is showing');
 
-  this.$('.md-dialog-container').mousedown().mouseup().click();
-  assert.ok(this.$('md-dialog').length, 'dialog is still showing');
+  $('.md-dialog-container').mousedown().mouseup().click();
+  assert.ok($('md-dialog').length, 'dialog is still showing');
 });
 
 test('dialog shouldn\'t swallow click events', function(assert) {
@@ -160,25 +170,23 @@ test('dialog shouldn\'t swallow click events', function(assert) {
   });
 
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{#paper-dialog clickOutsideToClose=true}}
       <button id="the-button">Go somewhere</button>
     {{/paper-dialog}}
   `);
 
-  assert.ok(this.$('md-dialog').length, 'dialog is showing');
+  assert.ok($('md-dialog').length, 'dialog is showing');
 
-  this.$('#the-button').mousedown().mouseup().click();
-  assert.ok(this.$('md-dialog').length, 'dialog is still showing');
+  $('#the-button').mousedown().mouseup().click();
+  assert.ok($('md-dialog').length, 'dialog is still showing');
 });
 
 test('has opt-in support for fullscreen at responsive breakpoint', function(assert) {
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{paper-dialog fullscreen=true}}
   `);
 
-  assert.ok(this.$('md-dialog').hasClass('md-dialog-fullscreen'), 'has class for fullscreen');
+  assert.ok($('md-dialog').hasClass('md-dialog-fullscreen'), 'has class for fullscreen');
 });
 
 test('pressing escape triggers close action', function(assert) {
@@ -192,17 +200,16 @@ test('pressing escape triggers close action', function(assert) {
   });
 
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{#if showDialog}}
       {{paper-dialog onClose=closeDialog}}
     {{/if}}
   `);
 
-  assert.ok(this.$('md-dialog'), 'dialog is showing');
+  assert.ok($('md-dialog'), 'dialog is showing');
 
   let event = new $.Event('keydown');
   event.keyCode = 27;
-  this.$('md-dialog').trigger(event);
+  $('md-dialog').trigger(event);
 
 });
 
@@ -214,7 +221,6 @@ test('opening gives focus', function(assert) {
   });
 
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{#if showDialog}}
       {{#paper-dialog onClose=closeDialog origin="#theorigin"}}
         {{#paper-dialog-actions}}
@@ -227,19 +233,19 @@ test('opening gives focus', function(assert) {
     </button>
   `);
 
-  this.$('#theorigin').focus();
-  assert.equal(document.activeElement, this.$('#theorigin').get(0));
-  this.$('#theorigin').click();
+  $('#theorigin').focus();
+  assert.equal(document.activeElement, $('#theorigin').get(0));
+  $('#theorigin').click();
 
   let done = assert.async();
 
   return wait().then(() => {
-    assert.equal(document.activeElement, this.$('#thedialogbutton').get(0));
+    assert.equal(document.activeElement, $('#thedialogbutton').get(0));
     this.set('showDialog', false);
 
     return wait();
   }).then(() => {
-    assert.equal(document.activeElement, this.$('#theorigin').get(0));
+    assert.equal(document.activeElement, $('#theorigin').get(0));
     done();
   });
 
@@ -247,20 +253,18 @@ test('opening gives focus', function(assert) {
 
 test('can specify dialog container classes', function(assert) {
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{paper-dialog dialogContainerClass="flex-50 my-dialog-container"}}
   `);
 
-  assert.ok(this.$('.md-dialog-container').hasClass('flex-50'), 'has flex-50 css class');
-  assert.ok(this.$('.md-dialog-container').hasClass('my-dialog-container'), 'has my-dialog-container css class');
+  assert.ok($('.md-dialog-container').hasClass('flex-50'), 'has flex-50 css class');
+  assert.ok($('.md-dialog-container').hasClass('my-dialog-container'), 'has my-dialog-container css class');
 });
 
 test('can specify dialog css classes', function(assert) {
   this.render(hbs`
-    <div id="paper-wormhole"></div>
     {{paper-dialog class="flex-50 my-dialog-inner"}}
   `);
 
-  assert.ok(this.$('md-dialog').hasClass('flex-50'), 'has flex-50 css class');
-  assert.ok(this.$('md-dialog').hasClass('my-dialog-inner'), 'has my-dialog-inner css class');
+  assert.ok($('md-dialog').hasClass('flex-50'), 'has flex-50 css class');
+  assert.ok($('md-dialog').hasClass('my-dialog-inner'), 'has my-dialog-inner css class');
 });