+++ tests/unit/utils/clamp-test.js
@@ -1,0 +1,14 @@
+import clamp from 'dummy/utils/clamp';
+import { module, test } from 'qunit';
+
+module('Unit | Utility | clamp');
+
+test('it clamps a number to minimum value', function(assert) {
+  let result = clamp(-10, 10, 20);
+  assert.equal(result, 10);
+});
+
+test('it clamps a number to maximum value', function(assert) {
+  let result = clamp(999999, 10, 20);
+  assert.equal(result, 20);
+});